transforms transformation_config inputs to spark dataset schema
