+++ .scalafmt.conf
@@ -1,0 +1,4 @@
+version = "3.7.17"
+maxColumn = 120
+align.preset = most
+continuationIndent.callSite = 2
